Move generic reader to separate file
